[reafactor/#44] refactor component & preview
